Fixed build. Introduced root device.
