+++ mips/malta.c
@@ -168,6 +168,8 @@ static void thread_bootstrap() {
   PCPU_SET(curthread, td);
 }
 
+extern void rootdev_init();
+
 void platform_init(int argc, char **argv, char **envp, unsigned memsize) {
   /* clear BSS section */
   bzero(__bss, __ebss - __bss);
@@ -179,7 +181,7 @@ void platform_init(int argc, char **argv, char **envp, unsigned memsize) {
   cpu_init();
   tlb_init();
   intr_init();
-  mips_intr_init();
+  rootdev_init();
   pm_bootstrap(memsize);
   kmem_bootstrap();
   sleepq_init();
